Profession Priority
Changed the priority that professions fill up the slots with. Set leadership to last so that it may be set to fill up all remaining slots. Increased the priority of Alchemy, Artificing and Weaponsmithing with Alchemy on top.

Included a few elements for compatibility with the changes included in: "Enabling alternate task lists for the same type"