Profession Priority
Changed the priority that professions fill up the slots with. Set leadership to last so that it may be set to fill up all remaining slots. Increased the priority of Alchemy, Artificing and Weaponsmithing with Alchemy on top.

Made it so that the priority list is smaller so that it can be easily modified if ever necessary. Position of the task lists themselves no longer matters.