Various updates
-- If searchForTask couldn't find the task and returned false,
createNextTask went into an infinite recursive loop, until it got a
stack overflow.  This happened, e.g., when there was no gold left to
buy resources.

Because of apparent limitations of JavaScript, the solution here was
the massive hack (imho) of keeping a global array of taskListNames,
populated if no task for that task list (profession name) could be
started, so the loop will move on to the next task list (profession)
to fill the slot.

This failed profession list is reset when the script switches
characters (or when it's reloaded/restarted).

-- If there's no task list at a given level, look for a lower level
task list, and use that.

-- If you made multiple task lists in the same profession, they'd step
on each other.  This was happening in Leadership, but I really wanted
to do it in Weaponsmithing, in order to make the best use of the
crafted resources.  The solution is, in effect, to assign specific
slots to task lists, in the priority order in which they're defined.
One consequece of that is that if the tasks get "mixed up" (e.g., by
manual assignment), a higher priority task might not start until it's
assigned slot opens up.  No slots will go unused, however, because if
a slot can't be filled by it's assigned, high priority task, it will
be filled by a lower priority task -- usually Leadership, if you keep
9 slots assigned to Leadership.

-- Make UI box a little bigger (to accommodate more task lists)

-- Option to purchase Porridge (which is the most time-efficient way
to advance Leadership at low levels).

-- Left notes on the structure of several objects in place, for future
reference.

-- Leadership task list reworked to take advantage of multi-task
sequences that are AD-efficient.  It's probably not complete/perfect
as yet, but improved.  I left my notes in place for future reference.

-- Reworked Weaponsmithing to account for the several different
classes of things that can be made.  It's generally more efficient to
make bows (which use wood) and one of the
daggers/greatswords/longswords (which use metal).