Update main.js
I don't know if it's only my problem, but sometimes (couple of times a day) script go into infinite loop, i.e. browser shows selected task and log shows that script try to push Start button. It is possible that this is the result of delays or poor hardware.
This addition count how many times IN A ROW particular character try to start particular task.
Because we have only 9 slots, i  think 10 attempt means that something goes wrong.
In such a situation this addition restart gateway (as is every 3 hours).