Update main.js
Profession asset updater added per character per profession.

Selecting "yes" and selecting a number of slots will allocate that number of slots to the production of new white workers for that profession.  It will fall back to the available professions for the current level and has error trapping in case of no "U" line in a profile.