Add char number to Resource Tracker
This should help accountability somehow. 

This way people that want to set different number of characters to produce different types of goods can easily see up to what character they have to copy the settings to (because not everyone has named their farm account characters as "FancyPants 1", "FancyPants2",.... "FancyPants50"), etc...

Also, I needed a place to track up to what character I've done what and it's easier to remember the number that the name for me.

The resource tracker might not be the best place for this or this should be on all tracking tabs so feel free to move/improve.